CR-997 Refactoring to just have sendEvent method
