+++ src/main/java/uk/gov/ons/ctp/common/event/persistence/EventPersistence.java
@@ -7,8 +7,6 @@ import uk.gov.ons.ctp.common.event.model.GenericEvent;
 
 public interface EventPersistence {
 
-  boolean isFirestorePersistenceSupported();
-
   void persistEvent(EventType eventType, RoutingKey routingKey, GenericEvent genericEvent)
       throws CTPException;
 }
